Split `AppState` to a standalone component; put data fetching hooks in one file.
